pyproject.toml
[project]
name = "medical-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ffmpeg>=1.4",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "langchain-openai>=0.3.18",
    "langchain-pinecone>=0.2.6",
    "langgraph>=0.4.3",
    "langsmith>=0.3.42",
    "pandas>=2.2.3",
    "pinecone>=6.0.2",
    "pydantic==2.11.5",
    "pydub>=0.25.1",
    "pypdf>=5.5.0",
    "pyprojroot>=0.3.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "streamlit-chat-widget>=0.2.0",
    "streamlit-extras>=0.7.1",
    "streamlit-float>=0.3.5",
    "uvicorn>=0.34.3",
]
